Fix LP history enqueue to respect Celery eager mode in tests.
Move catalog type-id helpers out of loyalty_store so market.tasks can import them without a cycle, and call apply_async instead of send_task-by-name.
